Refactor tests that get templates and move them all into one file
